backend/requirements.txt
django
djangorestframework
django-cors-headers
gunicorn
psycopg
psycopg2
djangorestframework_simplejwt